Refactored the Project Structure
cSVReaderUtil : load the csv file and map in to UserBean

UserBean : Bean class for csv mapping and will be used by Dataproviders
